Refactor UI styles and update project references
Added Aero-like brushes and button styles to ClassicStyle.xaml for improved Windows 7 look. MainWindow.xaml now uses standard window chrome and dynamic background brush. Updated project to use System.Management NuGet package for .NET Core/.NET 6 compatibility. Added AssemblyInfo.cs to suppress CLS compliance warnings. Refactored register access in ExoticFilesystemManager for long addressing. Fixed instruction dispatch type casting for x86. Minor code cleanup in CortexA15Cpu and HardwareHealthProbe.

+++ InstructionDispatcher.cs
@@ -28,7 +28,8 @@ namespace ProcessorEmulator.Emulation
                         break;
                     case "x86":
                     case "x64":
-                        x86Emu.DispatchInstruction(instruction, targetArch);
+                        // X86 emulator APIs use int for instructions; cast from uint safely here
+                        x86Emu.DispatchInstruction((int)instruction, targetArch);
                         break;
                     case "ARM":
                     case "ARM64":
